[Finish]完成

+++ app/controllers/orders_controller.rb
@@ -29,18 +29,21 @@ class OrdersController < ApplicationController
     @order = Order.new(order_params)
     @order.customer_id = current_customer.id
     @order.status = 0
-    @order.save
-    @cart_items = CartItem.where(customer_id: current_customer.id)
-      @cart_items.each do |cart_item|
-        @order_detail = @order.order_details.new
-        @order_detail.item_id = cart_item.item.id
-        @order_detail.amount = cart_item.amount
-        @order_detail.purchase_price = cart_item.item.price
-        @order_detail.product_status = 0
-        @order_detail.save
-      end
-    current_customer.cart_items.destroy_all
-    redirect_to orders_thanks_path
+    if @order.save
+      @cart_items = CartItem.where(customer_id: current_customer.id)
+        @cart_items.each do |cart_item|
+          @order_detail = @order.order_details.new
+          @order_detail.item_id = cart_item.item.id
+          @order_detail.amount = cart_item.amount
+          @order_detail.purchase_price = cart_item.item.price
+          @order_detail.product_status = 0
+          @order_detail.save
+        end
+      current_customer.cart_items.destroy_all
+      redirect_to orders_thanks_path
+    else
+      render :new
+    end
   end
 
   def index
